moved http_server stuff to main and created custom Flask thingy
